+++ README.md
@@ -1,35 +1,3 @@
-<<<<<<< HEAD
 # api
 
 Generated with express-generator
-=======
-# SmartBard
-SmartBard is a bulletin board web application for Overbrook School of the Blind designed to help those visually impaired to easily access items that would be inaccessible on a physical bulletin board.
-
-## Contents
-- [Contents](#contents)
-- [About](#about)
-- [Documentation](#documentation)
-- [Team](#team)
-- [License](#license)
-- [Supported-Browsers](#supported-browsers)
-
-## About
-Senior Project Design 
-
-## Documentation
-
-## Team
-- Peter Chong
-- Brandon Ly
-- Jarod Durkin
-- Sandy Yang
-- An Tran
-- Ricky Taing
-
-## License
-See [`License`](LICENSE.md).
-
-## Supported Browsers
-The platform is only tested in Mozilla Firefox, Google Chrome, and Safari. Other browsers might not render all pages correctly.
->>>>>>> Update README.md
