--- api/README.md
@@ -1,3 +1,0 @@
-# api
-
-Generated with express-generator
