--- api/routes/users.js
@@ -1,17 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-// const {
-//   getUsers,
-//   getUserById
-// } = require('../db/db-user-interface');
-
-/* GET users listing. */
-router.get('/', function(req, res, next) {
-  res.send('respond with a resource');
-});
-
-// router.route('/').get(getUsers);
-// router.route('/:id').get(getUserById);
-
-module.exports = router;
